Add DetectActive helper
